Rename `@require` to `@required`
